requirements.txt
# File: requirements.txt
requests
python-dotenv
